Upgrade to new elm-test
I copied the Model from the example since it was used in the tests and there is no longer a way to import another src directory for just testing purposes. Not sure how to handle the date situation. That parser is a poor example of good date handling, but it shows a concept.